HLM-5118 code refractoring, changing service name to resource-estimation service, adding comments
