+++ health-services/resource-estimation-service/pom.xml
@@ -1,15 +1,15 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.egov</groupId>
-    <artifactId>file-processor-utility</artifactId>
+    <artifactId>resource-estimation-service</artifactId>
     <packaging>jar</packaging>
     <name>file-processor-utility</name>
     <version>1.0.0</version>
     <properties>
     <java.version>17</java.version>
         <geotools.version>32-SNAPSHOT</geotools.version>
-    <maven.compiler.source>${java.version}</maven.compiler.source>
-    <maven.compiler.target>${java.version}</maven.compiler.target>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
 </properties>
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -97,19 +97,9 @@
             <version>4.1.2</version>
         </dependency>
         <dependency>
-            <groupId>org.apache.tika</groupId>
-            <artifactId>tika-core</artifactId>
-            <version>2.9.2</version> <!-- Use the latest version -->
-        </dependency>
-        <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-test</artifactId>
             <version>5.3.10</version> <!-- Use the appropriate version -->
-        </dependency>
-        <dependency>
-            <groupId>com.linuxense</groupId>
-            <artifactId>javadbf</artifactId>
-            <version>0.4.0</version>
         </dependency>
