+++ health-services/resource-estimation-service/src/main/java/org/egov/processor/web/models/Resource.java
