+++ health-services/resource-estimation-service/src/main/resources/application.properties
