+++ health-services/resource-estimation-service/src/main/resources/db/Dockerfile
